REFACTOR : 채팅 관련 리팩토링
